refactor(library): extract processLibraryFiles into separate FileAsBookProcessor class
